database migration and model update

+++ database/migrations/2025_05_07_170946_create_stock_items_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('stock_id')->nullable();
             $table->unsignedBigInteger('output_item_id')->nullable();
-            $table->double('selling_price')->nullable();
+            $table->decimal('selling_price', 15, 2)->nullable();
             $table->integer('available_qty')->nullable();
             $table->timestamps();
         // 'stock_id',
